--- chia-client/src/error.rs
@@ -1,5 +1,0 @@
-mod request_error;
-mod send_error;
-
-pub use request_error::*;
-pub use send_error::*;
