+++ chia-client/src/lib.rs
@@ -1,9 +1,3 @@
-mod error;
 mod peer;
-mod peer_event;
-mod peer_message;
 
-pub use error::*;
 pub use peer::*;
-pub use peer_event::*;
-pub use peer_message::*;
